requirements.txt
anthropic>=0.18.0
requests>=2.28.0
python-dateutil>=2.8.0
pytz>=2023.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
